core_test.py still fixing randint issue

+++ tests/core_test.py
@@ -36,7 +36,7 @@ class SpectrumConstructorTests(unittest.TestCase):
     """Test Spectrum.__init__()."""
 
     test_length = 256
-    data = int(np.random.randint(0, high=1e4, size=test_length))
+    data = np.random.randint(0, high=1e4, size=test_length).astype(np.int)
     test_gain = 8.23
     energy_edges = np.arange(test_length + 1) * test_gain
